Improve version page (add permission + tasklist for each version)

+++ src/version.php
@@ -1,3 +1,4 @@
+
 <?php
 define( 'ACTIVE_MENU', 'proj');
 include_once 'core/header.php';
@@ -8,6 +9,63 @@ $selectedProject = null;
 $projDA = new ProjectDA();
 $permDA = new PermissionDA();
 $alerts = "";
+
+if (isset($_GET["list"])) {
+	include_once 'core/logic/taskDA.php';
+	$taskDA = new TaskDA();
+	$currentVersion = $projDA->getVersionById($_GET["list"])->fetch_assoc();
+	$selectedProject = $projDA->getProject($currentVersion["project_id"]);
+	echo '<div id="main">
+			<ol class="breadcrumb">
+			  <li><a href="index.php">Home</a></li>
+			  <li><a href="project.php?p='. $selectedProject["id"] .'">Project '.$selectedProject["name"].'</a></li>
+			  <li><a href="version.php?p='. $selectedProject["id"] .'">Versions</a></li>
+			  <li class="active">Changelog '.$currentVersion["name"].'</li>
+			</ol>
+		
+			<h1>Changelog '.$selectedProject["name"].' '.$currentVersion["name"].'</h1>
+			<div>';
+	$tasks = $taskDA->getTasksByVersionID($_GET["list"]);
+	if ($tasks != null) {
+		$areTaskAvailable = false;
+		while ($oneTask = $tasks->fetch_assoc()) {
+			$component = $oneTask["componentName"];
+			if ($component != "") {
+				$component = '<span class="text-success pull-center" style="margin-left: 20px;">Component: '.$oneTask["componentName"].'</span>';
+			}
+			$assignee = $oneTask["assigneePrename"];
+			if ($assignee == "") {
+				$assignee = "none";
+			}
+			echo '<a href="task.php?t='.$oneTask["id"].'" class="list-group-item">
+							<div>
+				              <b>'.$oneTask["key"].'-'.$oneTask["id"].'</b>: '.
+							              $oneTask["summary"]. '' .
+							              ' <span class="badge pull-right" style="background-color: '.$oneTask["color"].'">'.$oneTask["name"].'</span>
+	  		                </div>
+							<div style="margin-left: 10px;"><em>
+	  							<span class="text-info">Assignee: '.$assignee.' '.$oneTask["assigneeSurname"].'</span>
+	  							'.$component.'
+						    </em></div>
+						  </a>';
+			$areTaskAvailable = true;
+		}
+		if (!$areTaskAvailable) {
+			echo '<span class="list-group-item">No Tasks founded...</span>';
+		}
+	}
+	else {
+		$permDA->echoPermissionDeniedAndDie();
+	}
+	echo '</div></div>';
+	include 'core/footer.php';
+	die();
+}
+
+if (!isset($_GET["p"])) {
+	$permDA->echoPermissionDeniedAndDie();
+}
+$writeAllowed = $permDA->isAdminOnProjectAllowed($_GET["p"]);
 
 if (isset($_GET["p"])) {
 	$selectedProject = $projDA->getProject($_GET["p"]);
@@ -21,7 +79,7 @@ if (!$permDA->isReadOnProjectAllowed($_GET["p"])) {
 	$permDA->echoPermissionDeniedAndDie();
 }
 
-if (isset($_POST["newVersion"])) {
+if (isset($_POST["newVersion"]) && $writeAllowed) {
 	$isReleased = 0;
 	if (isset($_POST["released"])) {
 		$isReleased = 1;
@@ -33,7 +91,7 @@ if (isset($_POST["newVersion"])) {
 	$projDA->createNewVersionForProject($_POST["p"], $_POST["versionname"], $_POST["description"], $isReleased, $date);
 }
 
-if (isset($_POST["editVersion"])) {
+if (isset($_POST["editVersion"]) && $writeAllowed) {
 	$isReleased = 0;
 	if (isset($_POST["released"])) {
 		$isReleased = 1;
@@ -45,7 +103,7 @@ if (isset($_POST["editVersion"])) {
 	$projDA->editVersion($_POST["versionId"], $_POST["versionname"], $_POST["description"], $isReleased, $date);
 }
 
-if (isset($_GET["del"])) {
+if (isset($_GET["del"]) && $writeAllowed) {
 	$projDA->deleteVersion($_GET["versionId"]);
 }
 
@@ -60,6 +118,7 @@ if (isset($_GET["del"])) {
 	
 	<h1>Versions Overview</h1>
 	<div>
+		<?php if ($writeAllowed) {?>
 		<button class="btn btn-default" data-toggle="modal" data-target="#newVersionModal">
 		  <i class="fa fa-plus-square"></i> New Version...
 		</button>
@@ -102,6 +161,7 @@ if (isset($_GET["del"])) {
 		    </div>
 		  </div>
 		</div>
+		<?php }?>
 	</div>
 	<h2><i class="fa fa-suitcase"></i> Unreleased Versions</h2>
 	<div class="list-group">
@@ -124,51 +184,61 @@ if (isset($_GET["del"])) {
 		if ($versionDate != "") {
 			$releaseDateString = 'Release Day: '.$oneVersion["doneDate"];
 		}
-		echo '<a href="#" class="list-group-item" data-toggle="modal" data-target="#version'.$oneVersion["id"].'Modal">
-			  	<i class="fa fa-tag"></i> <span style="font-weight: bold; width: 200px; display: inline-block;">
-					'.$oneVersion["name"].'</span> '.$oneVersion["description"].' <span style="float:right; color: '.$dateColor.'; 
-					font-style:italic;">'.$releaseDateString.'</span>
-			  </a>';
-		echo '<div class="modal fade" id="version'.$oneVersion["id"].'Modal" tabindex="-1" role="dialog" aria-labelledby="myModalLabel" aria-hidden="true">
-				  <div class="modal-dialog">
-				    <div class="modal-content">
-				      <form action="" method="post">
-	  			          <input type="hidden" name="editVersion" value="true" />
-						  <input type="hidden" name="versionId" value="'.$oneVersion["id"].'" />
-					      <div class="modal-header">
-					        <button type="button" class="close" data-dismiss="modal" aria-hidden="true">&times;</button>
-					        <h4 class="modal-title" id="myModalLabel">Version '.$oneVersion["name"].'</h4>
-					      </div>
-					      <div class="modal-body">
-				      		<input type="hidden" name="versionId" value="'.$oneVersion["id"].'" />
-					        <table width="100%">
-					        	<tr>
-					        		<td>Name:</td>
-					        		<td><input type="text" class="form-control" name="versionname" value="'.$oneVersion["name"].'" /></td>
-					        	</tr>
-					        	<tr>
-					        		<td>Description:</td>
-					        		<td><input type="text" class="form-control" name="description" value="'.$oneVersion["description"].'" /></td>
-					        	</tr>
-					        	<tr>
-					        		<td>Release Day:</td>
-					        		<td><input type="date" class="form-control" name="releaseDay" value="'.$oneVersion["doneDate"].'" /></td>
-					        	</tr>
-					        	<tr>
-					        		<td>Released?</td>
-					        		<td><input type="checkbox" class="form-control" name="released" /></td>
-					        	</tr>
-					        </table>
-					      </div>
-					      <div class="modal-footer">
-			        		<a href="?p='.$_GET["p"].'&versionId='.$oneVersion["id"].'&del=true" class="btn btn-danger"><i class="fa fa-trash-o"></i> Delete</a>
-					        <button type="button" class="btn btn-default" data-dismiss="modal">Close</button>
-					        <button type="submit" class="btn btn-primary">Save changes</button>
-					      </div>
-				      </form>
-				    </div>
-				  </div>
-				</div>';
+		if ($writeAllowed) {
+			echo '<a href="#" class="list-group-item" data-toggle="modal" data-target="#version'.$oneVersion["id"].'Modal">
+				  	<i class="fa fa-tag"></i> <span style="font-weight: bold; width: 200px; display: inline-block;">
+						'.$oneVersion["name"].'</span> '.$oneVersion["description"].' <span style="float:right; color: '.$dateColor.'; 
+						font-style:italic;">'.$releaseDateString.'</span>
+				  </a>';
+			echo '<div class="modal fade" id="version'.$oneVersion["id"].'Modal" tabindex="-1" role="dialog" aria-labelledby="myModalLabel" aria-hidden="true">
+					  <div class="modal-dialog">
+					    <div class="modal-content">
+					      <form action="" method="post">
+		  			          <input type="hidden" name="editVersion" value="true" />
+							  <input type="hidden" name="versionId" value="'.$oneVersion["id"].'" />
+						      <div class="modal-header">
+						        <button type="button" class="close" data-dismiss="modal" aria-hidden="true">&times;</button>
+						        <h4 class="modal-title" id="myModalLabel">Version '.$oneVersion["name"].'</h4>
+						      </div>
+						      <div class="modal-body">
+					      		<input type="hidden" name="versionId" value="'.$oneVersion["id"].'" />
+						        <table width="100%">
+						        	<tr>
+						        		<td>Name:</td>
+						        		<td><input type="text" class="form-control" name="versionname" value="'.$oneVersion["name"].'" /></td>
+						        	</tr>
+						        	<tr>
+						        		<td>Description:</td>
+						        		<td><input type="text" class="form-control" name="description" value="'.$oneVersion["description"].'" /></td>
+						        	</tr>
+						        	<tr>
+						        		<td>Release Day:</td>
+						        		<td><input type="date" class="form-control" name="releaseDay" value="'.$oneVersion["doneDate"].'" /></td>
+						        	</tr>
+						        	<tr>
+						        		<td>Released?</td>
+						        		<td><input type="checkbox" class="form-control" name="released" /></td>
+						        	</tr>
+						        </table>
+						      </div>
+						      <div class="modal-footer">
+								<a href="?list='.$oneVersion["id"].'" class="btn btn-default">View Changelog</a>
+				        		<a href="?p='.$_GET["p"].'&versionId='.$oneVersion["id"].'&del=true" class="btn btn-danger" onclick="return askBeforSending(\'Delete this Version forever?\')"><i class="fa fa-trash-o"></i> Delete</a>
+						        <button type="button" class="btn btn-default" data-dismiss="modal">Close</button>
+						        <button type="submit" class="btn btn-primary">Save changes</button>
+						      </div>
+					      </form>
+					    </div>
+					  </div>
+					</div>';
+		}
+		else {
+			echo '<a href="?list='.$oneVersion["id"].'" class="list-group-item">
+				  	<i class="fa fa-tag"></i> <span style="font-weight: bold; width: 200px; display: inline-block;">
+						'.$oneVersion["name"].'</span> '.$oneVersion["description"].' <span style="float:right; color: '.$dateColor.';
+						font-style:italic;">'.$releaseDateString.'</span>
+				  </a>';
+		}
 	} 
 	if (!$existsVersions) {
 		echo '<span class="list-group-item"><i class="fa fa-tag"></i> Nothing to display... </span>';
@@ -191,51 +261,61 @@ if (isset($_GET["del"])) {
 			if ($oneVersion["doneDate"] != "") {
 				$releaseDateString = 'Release Day: '.$oneVersion["doneDate"];
 			}
-			echo '<a href="#" class="list-group-item" data-toggle="modal" data-target="#version'.$oneVersion["id"].'Modal">
+			if ($writeAllowed) {
+				echo '<a href="#" class="list-group-item" data-toggle="modal" data-target="#version'.$oneVersion["id"].'Modal">
+					  	<i class="fa fa-tag"></i> <span style="font-weight: bold; width: 200px; display: inline-block;">
+							'.$oneVersion["name"].'</span> '.$oneVersion["description"].' <span style="float:right; color: #2FBD47; 
+							font-style:italic;">'.$releaseDateString.'</span>
+					  </a>';
+				echo '<div class="modal fade" id="version'.$oneVersion["id"].'Modal" tabindex="-1" role="dialog" aria-labelledby="myModalLabel" aria-hidden="true">
+					  <div class="modal-dialog">
+					    <div class="modal-content">
+					      <form action="" method="post">
+							  <input type="hidden" name="editVersion" value="true" />
+							  <input type="hidden" name="versionId" value="'.$oneVersion["id"].'" />
+						      <div class="modal-header">
+						        <button type="button" class="close" data-dismiss="modal" aria-hidden="true">&times;</button>
+						        <h4 class="modal-title" id="myModalLabel">Version '.$oneVersion["name"].'</h4>
+						      </div>
+						      <div class="modal-body">
+					      		<input type="hidden" name="versionId" value="'.$oneVersion["id"].'" />
+						        <table width="100%">
+						        	<tr>
+						        		<td>Name:</td>
+						        		<td><input type="text" class="form-control" name="versionname" value="'.$oneVersion["name"].'" /></td>
+						        	</tr>
+						        	<tr>
+						        		<td>Description:</td>
+						        		<td><input type="text" class="form-control" name="description" value="'.$oneVersion["description"].'" /></td>
+						        	</tr>
+						        	<tr>
+						        		<td>Release Day:</td>
+						        		<td><input type="date" class="form-control" name="releaseDay" value="'.$oneVersion["doneDate"].'" /></td>
+						        	</tr>
+						        	<tr>
+						        		<td>Released?</td>
+						        		<td><input type="checkbox" class="form-control" name="released" checked="checked" /></td>
+						        	</tr>
+						        </table>
+						      </div>
+						      <div class="modal-footer">
+								<a href="?list='.$oneVersion["id"].'" class="btn btn-default">View Changelog</a>
+			      				<a href="?p='.$_GET["p"].'&versionId='.$oneVersion["id"].'&del=true" class="btn btn-danger" onclick="return askBeforSending(\'Delete this Version forever?\')"><i class="fa fa-trash-o"></i> Delete</a>
+						        <button type="button" class="btn btn-default" data-dismiss="modal">Close</button>
+						        <button type="submit" class="btn btn-primary">Save changes</button>
+						      </div>
+					      </form>
+					    </div>
+					  </div>
+					</div>';
+			}
+			else {
+				echo '<a href="?list='.$oneVersion["id"].'" class="list-group-item">
 				  	<i class="fa fa-tag"></i> <span style="font-weight: bold; width: 200px; display: inline-block;">
-						'.$oneVersion["name"].'</span> '.$oneVersion["description"].' <span style="float:right; color: #2FBD47; 
+						'.$oneVersion["name"].'</span> '.$oneVersion["description"].' <span style="float:right; color: #2FBD47;
 						font-style:italic;">'.$releaseDateString.'</span>
 				  </a>';
-			echo '<div class="modal fade" id="version'.$oneVersion["id"].'Modal" tabindex="-1" role="dialog" aria-labelledby="myModalLabel" aria-hidden="true">
-				  <div class="modal-dialog">
-				    <div class="modal-content">
-				      <form action="" method="post">
-						  <input type="hidden" name="editVersion" value="true" />
-						  <input type="hidden" name="versionId" value="'.$oneVersion["id"].'" />
-					      <div class="modal-header">
-					        <button type="button" class="close" data-dismiss="modal" aria-hidden="true">&times;</button>
-					        <h4 class="modal-title" id="myModalLabel">Version '.$oneVersion["name"].'</h4>
-					      </div>
-					      <div class="modal-body">
-				      		<input type="hidden" name="versionId" value="'.$oneVersion["id"].'" />
-					        <table width="100%">
-					        	<tr>
-					        		<td>Name:</td>
-					        		<td><input type="text" class="form-control" name="versionname" value="'.$oneVersion["name"].'" /></td>
-					        	</tr>
-					        	<tr>
-					        		<td>Description:</td>
-					        		<td><input type="text" class="form-control" name="description" value="'.$oneVersion["description"].'" /></td>
-					        	</tr>
-					        	<tr>
-					        		<td>Release Day:</td>
-					        		<td><input type="date" class="form-control" name="releaseDay" value="'.$oneVersion["doneDate"].'" /></td>
-					        	</tr>
-					        	<tr>
-					        		<td>Released?</td>
-					        		<td><input type="checkbox" class="form-control" name="released" checked="checked" /></td>
-					        	</tr>
-					        </table>
-					      </div>
-					      <div class="modal-footer">
-		      				<a href="?p='.$_GET["p"].'&versionId='.$oneVersion["id"].'&del=true" class="btn btn-danger"><i class="fa fa-trash-o"></i> Delete</a>
-					        <button type="button" class="btn btn-default" data-dismiss="modal">Close</button>
-					        <button type="submit" class="btn btn-primary">Save changes</button>
-					      </div>
-				      </form>
-				    </div>
-				  </div>
-				</div>';
+			}
 		}
 		
 		if (!$existsVersions) {
